Refactor: Update project name and dependencies
This commit renames the project from "UnicoCaseAiVoiceChatApp" to "AiVoiceChatApp" and updates various dependencies to their latest versions.

Specific changes include:

- **Project Naming:**
    - Updated `applicationId` in `gradle/libs.versions.toml` from "com.ahmetsirim.unicocaseaivoicechatapp" to "com.ahmetsirim.aivoicechatapp".
    - Renamed main application class from `UnicoCaseAiVoiceChatApplication.kt` to `AiVoiceChatApplication.kt` and updated its package.
    - Renamed `MainActivity.kt` and updated its package.
    - Renamed `UnicoCaseAiVoiceChatNavHost.kt` to `AiVoiceChatNavHost.kt`.
    - Updated `app_name` string resource.
    - Updated theme names in `themes.xml` and `Theme.kt`.
    - Updated `rootProject.name` in `settings.gradle.kts`.
    - Renamed GitHub workflow file.
    - Updated namespace in `app/build.gradle.kts`.
    - Updated application and activity names in `AndroidManifest.xml`.
- **Dependency Updates in `gradle/libs.versions.toml`:**
    - `androidGradlePlugin`: "8.12.1" -> "8.13.0"
    - `kotlin`: "2.2.10" -> "2.2.20"
    - `lifecycleRuntimeKtx`: "2.9.2" -> "2.9.4"
    - `activityCompose`: "1.10.1" -> "1.11.0"
    - `ksp`: "2.2.10-2.0.2" -> "2.2.20-2.0.3"
    - `roomCompiler`: "2.7.2" -> "2.8.1"
    - `roomKtx`: "2.7.2" -> "2.8.1"
    - `hiltNavigation`: "1.2.0" -> "1.3.0"
    - `navigationComponent`: "2.9.3" -> "2.9.5"
    - `firebaseBom`: "34.1.0" -> "34.3.0"
    - `ktlint`: "11.6.1" -> "13.1.0"
    - Removed `jacoco` version.
- **Code Changes:**
    - Updated `hiltViewModel` import from `androidx.hilt.navigation.compose.hiltViewModel` to `androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel` in `HistoryContainer.kt`, `ChatContainer.kt`, and `SettingsContainer.kt`.
    - Changed default log tag in `LogExtensions.kt` from "UnicoCaseDebug" to "Debug".
    - Updated theme usage in `MainActivity.kt` and `InformationalDialog.kt`.
    - Updated typography variable name in `Type.kt`.
- **Build Configuration:**
    - Removed Jacoco test report task from `app/build.gradle.kts`.
- **File Deletion:**
    - Deleted `Jenkinsfile`.
