+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Unico Case Ai Voice Chat App</string>
+    <string name="app_name">Ai Voice Chat App</string>
 </resources>
