+++ core/designsystem/src/main/java/com/ahmetsirim/designsystem/theme/Type.kt
@@ -2,4 +2,4 @@ package com.ahmetsirim.designsystem.theme
 
 import androidx.compose.material3.Typography
 
-val unicoCaseAiVoiceChatAppTypography = Typography()
+val aiVoiceChatAppTypography = Typography()
