+++ feature/history/src/main/java/com/ahmetsirim/history/HistoryContainer.kt
@@ -2,7 +2,7 @@ package com.ahmetsirim.history
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 
 @Composable
