+++ feature/settings/src/main/java/com/ahmetsirim/settings/SettingsContainer.kt
@@ -2,12 +2,12 @@ package com.ahmetsirim.settings
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 
 @Composable
 fun SettingsContainer(
-    navigateUp: () -> Unit
+    navigateUp: () -> Unit,
 ) {
     val viewModel: SettingsViewModel = hiltViewModel()
     val uiState by viewModel.uiState.collectAsStateWithLifecycle()
