+++ gradle/libs.versions.toml
@@ -3,18 +3,18 @@
 compileSdk = "36"
 minSdk = "24"
 targetSdk = "36"
-applicationId = "com.ahmetsirim.unicocaseaivoicechatapp"
+applicationId = "com.ahmetsirim.aivoicechatapp"
 versionCode = "1"
 versionName = "1.0"
 
 # Android Gradle Plugin & Kotlin
-androidGradlePlugin = "8.12.1"
-kotlin = "2.2.10"
+androidGradlePlugin = "8.13.0"
+kotlin = "2.2.20"
 
 # AndroidX Core
 coreKtx = "1.17.0"
-lifecycleRuntimeKtx = "2.9.2"
-activityCompose = "1.10.1"
+lifecycleRuntimeKtx = "2.9.4"
+activityCompose = "1.11.0"
 
 # Compose Dependencies
 composeBom = "2025.08.00"
@@ -26,7 +26,7 @@ kotlinxSerializationJson = "1.9.0"
 kotlinxCoroutines = "1.10.2"
 
 # Kotlin Symbol Processor (KSP)
-ksp = "2.2.10-2.0.2"
+ksp = "2.2.20-2.0.3"
 
 # Retrofit & Moshi & OkHttp
 retrofit = "3.0.0"
@@ -34,19 +34,19 @@ moshi = "1.15.2"
 okhttpBom = "5.1.0"
 
 # Room Database
-roomCompiler = "2.7.2"
-roomKtx = "2.7.2"
+roomCompiler = "2.8.1"
+roomKtx = "2.8.1"
 
 # Hilt & Javax
 hilt = "2.57.1"
-hiltNavigation = "1.2.0"
+hiltNavigation = "1.3.0"
 javaxInject = "1"
 
 # Navigation Component
-navigationComponent = "2.9.3"
+navigationComponent = "2.9.5"
 
 # Firebase
-firebaseBom = "34.1.0"
+firebaseBom = "34.3.0"
 gmsGoogleServices = "4.4.3"
 crashlytics = "3.0.6"
 appdistribution = "5.1.1"
@@ -58,8 +58,7 @@ generativeAi = "0.9.0"
 timber = "5.0.1"
 secretsGradlePluginVersion = "2.0.1"
 identityJvm = "202411.1"
-ktlint = "11.6.1"
-jacoco = "0.8.8"
+ktlint = "13.1.0"
 
 [libraries]
 # AndroidX Core
